Implement formspec version chooser
